cheese-backend/requirements.txt
fastapi
uvicorn[standard]
openai
astrapy
python-dotenv
pydantic
